requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.7.0
pynostr==0.6.2
bech32==1.2.0
secp256k1==0.14.0
nostr-sdk==0.36.0
